requirements.txt
fastapi
uvicorn
openai
aio_pika
httpx
langchain.tools
requests
injector